requirements.txt
pynput
pystray
psutil
pywin32
pycaw
pyyaml
